review: split whimsy/index.tsx into 4 files + drop check-file-lines exemption + docstrings + memo cleanup
PR #137 review feedback (review id 4253711983):

1. KnowledgeUrlState (and its sibling KnowledgeNavigationHandlers)
   had no docstrings.  Added per-field TSDoc on both interfaces.
   The TSDoc rule technically already requires this; the gap is
   that we have no enforcement script for tsdoc-on-exports yet.
   Filing a follow-up to add one (instruction-only rules slip).

2. `openFilePath` useMemo justification: there was none, so I
   removed the memo.  It returns a primitive string, React Query
   compares query keys by value, so memoizing only added an
   allocation per render with no benefit.  Other memos in the hook
   (`segments`, `folderSegments`, `currentNode`, `openFileNode`,
   `crumbs`) all return arrays or object refs that downstream hooks
   use as deps; their memo justifications are now in inline
   comments naming the consumer that benefits.

3. whimsy/index.tsx file-length: the script's
   `EXEMPT_PATH_FRAGMENTS` carved it out as "pre-existing tech
   debt".  Per the no-pre-existing-excuse rule this PR ships
   alongside (#136), the exemption goes away and so does the
   underlying file.  Split into:

     - frontend/features/whimsy/config.ts (184 lines): types,
       bounds, validation, useWhimsyConfig hook.
     - frontend/features/whimsy/use-whimsy-tile.ts (93 lines):
       the CSS mask URL hook.
     - frontend/features/whimsy/WhimsySettingsCard.tsx (469
       lines): the settings card + WhimsyColorPicker + the three
       row sub-components from the previous split.
     - frontend/features/whimsy/index.tsx (29 lines): barrel
       re-exporting the public surface so the three external
       consumers (ChatView, SettingsLayout, AppearanceSection)
       don't change a line.

   Removed the exemption entry from check-file-lines.mjs.  All
   four whimsy files are under 500 lines on their own.

Verification:
  bun x @biomejs/biome check          -> 0 errors, 0 warnings
  bun x tsc --noEmit                   -> green
  node ../scripts/check-file-lines.mjs -> OK (whimsy no longer exempt)
  node ../scripts/check-nesting.mjs    -> OK
